fix(widgets): stop an explicit default cache extent panicking the build
Codex P2 on this PR, and a real crash rather than a style point.

`Viewport::build_render_object` asserted that `set_cache_extent` returns
`LAYOUT` for any explicitly-set cache extent. But `RenderViewport`'s own
default is 250.0 with `CacheExtentStyle::Pixel`, so
`Viewport::new(..).cache_extent(250.0, CacheExtentStyle::Pixel)` — a caller
simply being explicit about the default — makes `set_cache_extent` correctly
report `NONE`, and the assertion then panics in every debug and test build.

The assertion encoded an invariant that is not true: "explicit" does not imply
"different". Dropped, with a comment recording why the returned impact is
ignored here at all — this runs before the render object joins a tree, so there
is nothing to invalidate. `an_explicit_cache_extent_equal_to_the_default_builds_
without_panicking` fails with the assertion restored (verified).

Also from this review round (both Copilot):

- `tests/common/mod.rs` held a `#[test]`. Four integration crates do
  `mod common;`, so that one test was compiled and run four times. Moved to
  `tests/layer_inspection.rs`, which already owns layer-retention cases;
  `common` is helpers only again.
- `AnchoredBox::update_render_object`'s doc was a spliced fragment. Rewritten
  to say what it means: the anchor carries identity only, so a rebuild has
  nothing to invalidate, and adopting a replacement child is what schedules
  that child's layout.
